requirements.txt
uvicorn
openpyxl
fastapi
farm-haystack[inference]